Remove rounding from AnchoredInterval
Remove rounding from `AnchoredInterval{P}` (which was a special case when
`P` is a `Period`) and replace it with the pseudoconstructors `HE` and
`HB`. This fixes some problems associated arithmetic involving
`AnchoredInterval`s.

Also `intersect(AnchoredInterval, AnchoredInterval)` now returns an
`AnchoredInterval` instead of an `Interval` (but the tests are broken;
they'll be fixed once we change the value type parameter to a field).

Closes #1