Avoid extending Base.isopen
The `isopen` function defined here is very different and so it seems
best we define a new function rather than extend `Base.isopen`.
Doing this means we need to export `Intervals.isopen` to make this
change non-breaking however doing this results in:

```julia
WARNING: both Intervals and Base export "isopen"; uses of it in module Main must be qualified
ERROR: UndefVarError: isopen not defined
```

It then seems best to not export `Intervals.isopen` and also then not
export `Intervals.isclosed` to keep things consistent.