Fix AnchoredInterval scalar convert
In https://github.com/invenia/Intervals.jl/pull/101 I removed this
convert function as it was lossy and wasn't necessary. After that PR was
merged I discovered `convert(::Type{T}, ::Interval{T})` which defines
the conversion of an interval to a scalar when an interval only includes
a single value. It appears that the AnchoredInterval function was
supposed to do the same scalar conversion but it was implemented
incorrectly.